updated with refactor
